% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_paired_t_test.R
\name{equiv.paired.t.test}
\alias{equiv.paired.t.test}
\title{Equivalence Paired t-Test}
\usage{
equiv.paired.t.test(diff, alpha = 0.05, eps_sub, eps_std, eps_tol,
  power.out = FALSE)
}
\arguments{
\item{diff}{a numeric vector of the differences between paired observations from two treatment groups.}

\item{alpha}{a numeric value for the significance level, between 0 and 1 (default = 0.05).}

\item{eps_sub}{a numeric vector or numeric value (for a symmetric equivalence range about 0) on the scale of the variable.}

\item{eps_std}{a numeric vector or numeric value (for a symmetric equivalence range about 0) of the desired standard deviations to bound the equivalence region.}

\item{eps_tol}{a character string specifying the tolerance levels in Wellek (2010), must be either "strict" (0.25) or "liberal" (0.50).}

\item{power.out}{an optional logical vector to output power calculation.}
}
\value{
Returns a list with the following components:
\item{intype}{ the type of epsilon input from user.}
\item{t.stat}{ value of the t-statistic.}
\item{rej}{ TRUE value if null hypothesis rejected.}
\item{CI_std}{ equivalence confidence interval in terms of standard deviations.}
\item{CI_sub}{ equivalence confidence interval on the scale of the variable.}
\item{power}{ power of the test}
}
\description{
\code{equiv.paired.t.test} is a paired equivalence t-test for an asymptotically normal vector of individual differences taken from two treatment groups for symmetric equivalence ranges. Nonsymmetric equivalence ranges are also accepted, however p-values and confidence intervals will not be calculated at this time.
}
\details{
To set the equivalence range, enter one of the following variables for epsilon: \code{eps_sub} (preferred), \code{eps_std}, or \code{eps_tol}. Using knowledge of the data, \code{eps_sub} should be a substantive equivalence range on the scale of the variable. \code{eps_std} should specify an equivalence range of standard deviations. Otherwise, use the strict (default) and liberal epsilon tolerances, \code{eps_tol}, for the paired equivalence t-test as defined in Wellek (2010).
}
\examples{
diff = c(10.3,11.3,2,-6.1,6.2,6.8,3.7,-3.3,-3.6,-3.5,13.7,12.6)
equiv.paired.t.test(diff, alpha = .05, eps_std=.5)
}
\references{
Hartman, E., & Hidalgo, F. D. (2018). An Equivalence Approach to Balance and Placebo Tests. American Journal of Political Science, 62(4), 1000-1013.

Wellek, S. (2010). Testing statistical hypotheses of equivalence and noninferiority. Chapman and Hall/CRC.
}
\seealso{
\code{\link{equiv.t.test}}
}
