% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_binom_one_sided.R
\name{fisher.binom.one.sided}
\alias{fisher.binom.one.sided}
\title{Exact Fisher Binomial One-Sided Test for Two Samples}
\usage{
fisher.binom.one.sided(x, y, eps_std, eps_sub, eps_tol, alpha,
  power.out = FALSE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of binary responses from treatment group A.}

\item{y}{a (non-empty) numeric vector of binary responses from treatment group B.}

\item{eps_std}{a numeric value for epsilon to determine the 1-epsilon hypothesis.}

\item{eps_sub}{a numeric value representing the probability that is substantively considered to demonstrate nonequivalence between the two groups.}

\item{eps_tol}{a character string specifying the tolerance levels in Wellek (2010), must be either "strict" (0.41) or "liberal" (0.85).}

\item{alpha}{a numeric value for the significance level, between 0 and 1 (default = 0.05).}

\item{power.out}{an optional logical vector to output the power calculation.}
}
\value{
Returns a list with the following components:
\item{intype}{ the type of epsilon input from user.}
\item{odds.ratio}{ odds ratio of treatment groups A and B.}
\item{p}{ p-value of the test.}
\item{rej}{ 1 if null hypothesis is rejected.}
}
\description{
\code{fisher.binom.one.sided} is a one-sided equivalence Fisher type exact test for two binomial distributions with respect to the odds ratio. The alternative hypothesis is: the odds ratio is less than or equal to 1- \code{eps_std}.
}
\details{
To set the equivalence range, enter one of the following variables for epsilon: \code{eps_sub} (preferred), \code{eps_std}, or \code{eps_tol}. Using knowledge of the data, \code{eps_sub} should be the probability that is substantively considered to demonstrate nonequivalence between the two groups (Wellek 2010, p. 16). \code{eps_std} creates an upper bound of 1-epsilon for the odds ratio. Otherwise, use the strict (default) and liberal epsilon tolerances, \code{eps_tol}, for the one-sided Fisher-Binomial test as defined in Wellek (2010).
}
\examples{
# One-sided example from Wellek 2010, pg 176
A = c(rep(1, 98), rep(0, 8))
B = c(rep(1, 97), rep(0, 10))
fisher.binom.one.sided(A, B, eps_std=.5, alpha=0.05)
}
\references{
Hartman, E., & Hidalgo, F. D. (2018). An Equivalence Approach to Balance and Placebo Tests. American Journal of Political Science, 62(4), 1000-1013.

Wellek, S. (2010). Testing statistical hypotheses of equivalence and noninferiority. Chapman and Hall/CRC.
}
\seealso{
\code{\link{fisher.binom.two.sided}}
}
