% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{odds.ratio}
\alias{odds.ratio}
\title{Odds Ratio}
\usage{
odds.ratio(p1, p2)
}
\arguments{
\item{p1}{a (nonempty) numeric value of the average of the binary responses from treatment group A.}

\item{p2}{a (nonempty) numeric value of the average of the binary responses from treatment group B.}
}
\value{
Calculates the odds ratio using the outcomes from treatment groups A and B.
}
\description{
The odds ratio is an appropriate dissimilarity measure for two binomial samples (Wellek 2010). Therefore, this result helps determines our equivalence hypotheses for the \code{\link{fisher.binom.one.sided}} and \code{\link{fisher.binom.two.sided}} tests.
}
\examples{
# One-sided example (from Wellek 2010 pg 176)
A = c(rep(1, 98), rep(0, 8))
B = c(rep(1, 97), rep(0, 10))
odds.ratio(mean(A),mean(B))
}
\seealso{
\code{\link{fisher.binom.one.sided}}, \code{\link{fisher.binom.two.sided}}
}
