% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_equivfisherbinomial.R
\name{summary.equivfisherbinom}
\alias{summary.equivfisherbinom}
\title{Equivalence Fisher-Binomial Test Summary Function}
\usage{
\method{summary}{equivfisherbinom}(res)
}
\description{
\code{summary.equivfisherbinom} is a summary function for objects of the class \code{equivfisherbinom}. This summary function should be paired with the \code{fisher.binom.one.sided} or \code{fisher.binom.two.sided} functions.
}
\examples{
# One-sided example from Wellek 2010, pg 176
A = c(rep(1, 98), rep(0, 8))
B = c(rep(1, 97), rep(0, 10))
res = fisher.binom.one.sided(A, B, eps_std=.5, alpha=0.05)
summary(res)
}
\seealso{
\code{\link{fisher.binom.two.sided}}, \code{\link{fisher.binom.one.sided}}
}
