% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_equivpairedt.R
\name{summary.equivpairedt}
\alias{summary.equivpairedt}
\title{Equivalence Paired t-Test Summary Function}
\usage{
\method{summary}{equivpairedt}(res)
}
\description{
\code{summary.equivpairedt} is a summary function for objects of the class \code{equivpairedt}. This summary function should be paired with the \code{equiv.paired.t.test} function.
}
\examples{
# Wellek p 96
res = equiv.paired.t.test(diff.mu=.16, diff.sd= 3.99, n=23, alpha = .05, eps_std=.5)
summary(res)
}
\seealso{
\code{\link{equiv.paired.t.test}}
}
