% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tost.R
\name{summary.equivtost}
\alias{summary.equivtost}
\title{Equivalence Two-One-Sided Test Summary Function}
\usage{
\method{summary}{equivtost}(res)
}
\description{
\code{summary.equivtost} is a summary function for objects of the class \code{equivtost}. This summary function should be paired with the \code{tost} function.
}
\examples{
# Wellek p 124
x=c(10.3,11.3,2,-6.1,6.2,6.8,3.7,-3.3,-3.6,-3.5,13.7,12.6)
y=c(3.3,17.7,6.7,11.1,-5.8,6.9,5.8,3,6,3.5,18.7,9.6)
res=tost(x,y,eps_sub=10, alpha=.1)
summary(res)
}
\seealso{
\code{\link{equiv.tost}}
}
