% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_equivttest.R
\name{summary.equivttest}
\alias{summary.equivttest}
\title{Equivalence t-Test Summary Function}
\usage{
\method{summary}{equivttest}(res)
}
\description{
\code{summary.equivttest} is a summary function for objects of the class \code{equivttest}. This summary function should be paired with the \code{equiv.t.test} function.
}
\examples{
# Wellek p 124
x=c(10.3,11.3,2,-6.1,6.2,6.8,3.7,-3.3,-3.6,-3.5,13.7,12.6)
y=c(3.3,17.7,6.7,11.1,-5.8,6.9,5.8,3,6,3.5,18.7,9.6)
eps=c(.5,1)
res=equiv.t.test(x,y,eps_sub=eps)
summary(res)
}
\seealso{
\code{\link{equiv.t.test}}
}
