% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tost_ratio.R
\name{tost.ratio}
\alias{tost.ratio}
\title{Two-One-Sided Ratio Test (TOST Ratio)}
\usage{
tost.ratio(x, y, frac, alpha = 0.05, eps_sub, eps_std)
}
\arguments{
\item{x}{a (non-empty) numeric vector of treatment group A.}

\item{y}{a (non-empty) numeric vector of treatment group B.}

\item{frac}{a numeric value for the ratio of the population mean of treatment group A to the population mean of treatment group B.}

\item{alpha}{a numeric value for the confidence level, between 0 and 1 (default = 0.05).}

\item{eps_sub}{a numeric vector or numeric value (for a symmetric equivalence range about 0) on the scale of the variable.}

\item{eps_std}{a numeric vector or numeric value (for a symmetric equivalence range about 0) of the desired standard mean differences on the outcome using the pooled standard deviation.}
}
\value{
\item{rej}{ TRUE value if null hypothesis rejected.}
\item{tstat}{ the lower and upper test statistics. }
}
\description{
\code{tost.ratio} is a two-one-sided equivalence ratio test for the ratio of the population means of the two treatment groups. The alternative hypothesis tests whether the ratio of the treatment to control population means are within a given equivalence range.
}
\examples{
tost.ratio()
}
\references{
Berger, R. L., & Hsu, J. C. (1996). Bioequivalence trials, intersection-union tests and equivalence confidence sets. Statistical Science, 11(4), 283-319.

Hartman, E., & Hidalgo, F. D. (2018). An Equivalence Approach to Balance and Placebo Tests. American Journal of Political Science, 62(4), 1000-1013.
}
