% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tost.R
\name{tost}
\alias{tost}
\title{Two-One-Sided Test (TOST)}
\usage{
tost(x, y, eps_sub, eps_std, alpha = 0.05)
}
\arguments{
\item{x}{a (non-empty) numeric vector of treatment group A.}

\item{y}{a (non-empty) numeric vector of treatment group B.}

\item{eps_sub}{a numeric vector or numeric value (for a symmetric equivalence range about 0) on the scale of the variable.}

\item{eps_std}{a numeric vector or numeric value (for a symmetric equivalence range about 0) of the desired standard mean differences on the outcome using the pooled standard deviation.}

\item{alpha}{a numeric value for the confidence level, between 0 and 1 (default = 0.05).}
}
\value{
\item{rej}{ TRUE value if null hypothesis rejected.}
\item{tstat}{ the lower and upper test statistics. }
}
\description{
\code{tost} is a two-one-sided equivalence test for the raw mean difference. The alternative hypothesis tests whether the difference of the two treatment means are within a given equivalence range.
}
\details{
To set the equivalence range, enter one of the following variables for epsilon: \code{eps_sub} (preferred) or \code{eps_std}. Using knowledge of the data, \code{eps_sub} is the substantive mean value to bound the difference in means in the hypothesis. \code{eps_std} is the desired number of standard deviations that should be used to bound the difference in means. If no epsilon is specified, a default value of \code{eps_std} = 1 is used.
}
\examples{
tost()
}
\references{
Berger, R. L., & Hsu, J. C. (1996). Bioequivalence trials, intersection-union tests and equivalence confidence sets. Statistical Science, 11(4), 283-319.

Hartman, E., & Hidalgo, F. D. (2018). An Equivalence Approach to Balance and Placebo Tests. American Journal of Political Science, 62(4), 1000-1013.
}
